+++ src/main/java/frc/robot/RobotContainer.java
@@ -9,15 +9,15 @@ import frc.robot.commands.IndexerCommand;
 import frc.robot.commands.LimeLightIntakeCommand;
 import frc.robot.commands.SetElevatorPositionCommand;
 import frc.robot.commands.SetPivotPositionCommand;
-import frc.robot.commands.SetPivotPositionCommandTest;
 import frc.robot.commands.IntakeCommand;
-import frc.robot.commands.IntakeCommandTest;
 import frc.robot.commands.SetTurretPositionCommand;
-import frc.robot.commands.SetTurretPositionCommandTest;
 import frc.robot.commands.ShooterCommand;
-import frc.robot.commands.ShooterCommandTest;
 import frc.robot.commands.TestIndexerCommand;
 import frc.robot.commands.TestShooterCommand;
+import frc.robot.commands.tests.IntakeCommandTest;
+import frc.robot.commands.tests.SetPivotPositionCommandTest;
+import frc.robot.commands.tests.SetTurretPositionCommandTest;
+import frc.robot.commands.tests.ShooterCommandTest;
 import frc.robot.generated.TunerConstants;
 import frc.robot.sim.MechanismViewer;
 import frc.robot.subsystems.IndexerSubsystem;
