+++ src/main/java/frc/robot/commands/tests/SetPivotPositionCommandTest.java
@@ -1,7 +1,7 @@
-package frc.robot.commands;
+package frc.robot.commands.tests;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.commands.SetPivotPositionCommand;
 import frc.robot.subsystems.pivot.PivotSubsystem;
 
 /**
